some fixes.

+++ EDGAR-c-meta/job-1-getFinra-meta.sh
@@ -56,8 +56,11 @@ curl --header 'Accept:application/json, text/plain, */*' \
     --data @finra.content.json \
     -sSL "https://tools.finra.org/fa_api/search/funds" | tee |
     node ./node-process-finra.js | jq . >finra.tmp.json
-# need to put logic in to error out of the .tmp file emmpty.
 
+if [ ! -s "finra.tmp.json" ]; then
+    echo "Empty finra.tmp.jason file."
+    exit 1
+fi
 if ../bin/jsonDifferent.sh finra.tmp.json "$finraFile"; then
     cat finra.tmp.json >"$finraFile"
     echo $finraFile updated.
